refactor: Seperate model module
Moved model from core to a model module.
