+++ model/build.gradle
@@ -10,7 +10,6 @@ repositories {
 }
 
 dependencies {
-    implementation project(":model")
     testImplementation platform('org.junit:junit-bom:5.10.0')
     testImplementation 'org.junit.jupiter:junit-jupiter'
     implementation 'com.fasterxml.jackson.core:jackson-databind:2.19.0'
