+++ model/src/main/java/com/apighost/model/scenario/HTTPMethod.java
