+++ model/src/main/java/com/apighost/model/scenario/ProtocolType.java
