+++ model/src/main/java/com/apighost/model/scenario/result/ScenarioResult.java
